requirements.txt
pandas
streamlit
yt_dlp 
ffmpeg-python 
torchaudio 
transformers
pandas
numpy
# openai-whisper
